many new plugins, changes; see below
- Add vim-go for better golang support.
- Replace nginx-vim-syntax with nginx.vim.
- Add sslsecure.vim for warnings when using insecure ciphers.
- Disable syntastic.
- Add vim-speeddating for date formatting.
- Add ale for asynchronous, multi-language linting.
- Add (commented out) vim-illuminate to highlight identical words.

Also added some configuration to make work life easier: .md files are
always Markdown; .go, .c, and .h -> expandtab! Backups are enabled and
both backups and swap files are now placed in ~/.vim-runtime/ (to
prevent constant littering of directories with garbage files, and to
keep all the runtime writes on one partition.

Lastly, I added back the old hack to restore the cursor to its last
position. While this is not necessary on all systems, it is on some I use.